requirements.txt
python-telegram-bot==20.6
Flask==2.3.5
firebase-admin==6.1.0
requests==2.31.0
python-dotenv==1.0.0
numexpr==2.8.8
